Fix CodeQL's narrowing conversions in nineteen bounds checks
CodeQL flagged `phaseCapSim` in the simulation runner: a realm index and a
stage number that came out of SQLite as int64 were narrowed to `int` before
being compared against `len()`. On a 32-bit build that truncates, and a
truncated index passes a bounds check it should fail.

The same idiom was in eighteen other places — every realm, stage, room, quest
and boss-phase bounds check in the engine, all of them guarding a slice index
taken from the database. Widening the length to int64 is always safe, since
`len()` is non-negative and fits; narrowing the index is what is not. So all
nineteen now read `x >= int64(len(y))` rather than `int(x) >= len(y)`.

No behaviour change on a 64-bit build, which is every platform this ships to;
this is the check being correct rather than incidentally correct.

Go vet, build and the full Go suite pass; the engine playtest runs clean.

+++ go_core/internal/game/aptitude_actions.go
@@ -203,7 +203,7 @@ func loadEffectModifiers(conn *storage.Conn, userID, gameMinute int64, bundle Ap
 }
 
 func phaseCost(realms []worlddata.Realm, index, phase int64) (int64, error) {
-	if index < 0 || int(index) >= len(realms) {
+	if index < 0 || index >= int64(len(realms)) {
 		return 0, errors.New("realm index out of range")
 	}
 	costs := realms[index].PhaseCosts
