+++ go_core/internal/simulation/advanced_maintenance.go
@@ -638,11 +638,11 @@ func (r *Runner) phaseCapSim(realm, phase int64, body bool) int64 {
 	if body {
 		realms = r.Catalog.BodyRealms
 	}
-	if realm < 0 || int(realm) >= len(realms) || phase < 1 {
+	if realm < 0 || realm >= int64(len(realms)) || phase < 1 {
 		return math.MaxInt64
 	}
 	costs := realms[realm].PhaseCosts
-	if int(phase) > len(costs) {
+	if phase > int64(len(costs)) {
 		return math.MaxInt64
 	}
 	return costs[phase-1]
